prl-5304/move_citizen_response_docs_to_be_reviewed_folder: Refactored CaseApplicationResponseService
